Fixing ci
