+++ .github/workflows/main.yml
@@ -61,6 +61,7 @@ jobs:
         run: |
           brew install ccache
           brew install gcc
+          brew install gettext
 
       - name: show-gcc
         run: |
