+++ doc/whats_new/v0.8.rst
@@ -17,7 +17,10 @@ Changelog
     did not raise an error, and silently ran, assuming the missing-values were
     encoded as infinity value. This is now fixed, and the estimators will raise an
     ValueError if missing-values are encountered in ``X`` input array.
-    By `Adam Li`_ (:pr:`#254`)
+    By `Adam Li`_ (:pr:`#264`)
+- |Feature| Simulations in ``sktree.datasets.hyppo`` now throw a warning instead
+    of an error when the number of samples is less than the number of dimensions.
+    By `Sambit Panda`_ (:pr:`#279`)
 
 Code and Documentation Contributors
 -----------------------------------
