+++ examples/treeple/treeple_tutorial_2_1a_SA98_multiview.py
@@ -1,44 +1,47 @@
 """
-======================
-1-1a: Calculating S@98
-======================
+====================================
+Calculating S@98 with multiview data
+====================================
 """
 
 import matplotlib.pyplot as plt
 import numpy as np
+import pandas as pd
 import seaborn as sns
 from sklearn.metrics import roc_curve
 
 from sktree.datasets import make_trunk_classification
 from sktree.ensemble import HonestForestClassifier
-from sktree.stats import build_hyppo_oob_forest
+from sktree.stats import build_oob_forest
+from sktree.tree import MultiViewDecisionTreeClassifier
 
 sns.set(color_codes=True, style="white", context="talk", font_scale=1.5)
 PALETTE = sns.color_palette("Set1")
 sns.set_palette(PALETTE[1:5] + PALETTE[6:], n_colors=9)
 sns.set_style("white", {"axes.edgecolor": "#dddddd"})
-
 # %%
-# S@98
-# ----
+# S@98 with multiview data
+# ------------------------
 #
 # Sensitivity at 98% specificity (*S@98*) measures, namely, the true
 # positive rate (*TPR*) when the false positive rate (*FPR*) is at 98%.
 #
 # .. math:: S@r = \mathbb{P}[\eta(X) > T_r \mid Y=1]
 #
-# With a binary class simulation as an example, this tutorial will show
-# how to use ``treeple`` to calculate the statistic.
+# With a multiview binary class simulation as an example, this tutorial
+# will show how to use ``treeple`` to calculate the statistic with
+# multiview data. For data with a single feature set, you can check out
+# the simpler S@98 tutorial.
 
 # %%
-# Create a simulation with two gaussians
-# --------------------------------------
+# Create a two-dimensional simulation with gaussians
+# --------------------------------------------------
 
 
 # create a binary class simulation with two gaussians
 # 500 samples for each class, class zero is standard
 # gaussian, and class one has a mean at one
-X, y = make_trunk_classification(
+Z, y = make_trunk_classification(
     n_samples=1000,
     n_dim=1,
     mu_0=0,
@@ -47,22 +50,37 @@ X, y = make_trunk_classification(
     seed=1,
 )
 
+# class one has a mean at two for X
+X, y = make_trunk_classification(
+    n_samples=1000,
+    n_dim=1,
+    mu_0=0,
+    mu_1=2,
+    n_informative=1,
+    seed=2,
+)
+
+Z_X = np.hstack((Z, X))
+
+
+Z_X_y = np.hstack((Z_X, y.reshape(-1, 1)))
+Z_X_y = pd.DataFrame(Z_X_y, columns=["Z", "X", "y"])
+Z_X_y = Z_X_y.replace({"y": 0.0}, "Class Zero")
+Z_X_y = Z_X_y.replace({"y": 1.0}, "Class One")
 
 fig, ax = plt.subplots(figsize=(6, 6))
 fig.tight_layout()
 ax.tick_params(labelsize=15)
-
-# histogram plot the samples
-ax.hist(X[:500], bins=50, alpha=0.6, color=PALETTE[1], label="negative")
-ax.hist(X[500:], bins=50, alpha=0.3, color=PALETTE[0], label="positive")
-ax.set_xlabel("X", fontsize=15)
-ax.set_ylabel("Likelihood", fontsize=15)
+sns.scatterplot(data=Z_X_y, x="Z", y="X", hue="y", palette=PALETTE[:2][::-1], alpha=0.2)
+sns.kdeplot(data=Z_X_y, x="Z", y="X", hue="y", palette=PALETTE[:2][::-1], alpha=0.6)
+ax.set_ylabel("Variable Two", fontsize=15)
+ax.set_xlabel("Variable One", fontsize=15)
 plt.legend(frameon=False, fontsize=15)
-plt.show()
+
 
 # %%
-# Fit the model
-# -------------
+# Fit the model with X and Z
+# --------------------------
 
 
 # initialize the forest with 100 trees
@@ -72,11 +90,12 @@ est = HonestForestClassifier(
     max_features=0.3,
     bootstrap=True,
     stratify=True,
+    tree_estimator=MultiViewDecisionTreeClassifier(),
     random_state=1,
 )
 
 # fit the model and obtain the tree posteriors
-_, observe_proba = build_hyppo_oob_forest(est, X, y)
+_, observe_proba = build_oob_forest(est, Z_X, y)
 
 # generate forest posteriors for the two classes
 observe_proba = np.nanmean(observe_proba, axis=0)
