+++ sktree/stats/__init__.py
@@ -2,8 +2,8 @@ from .forestht import (
     FeatureImportanceForestClassifier,
     FeatureImportanceForestRegressor,
     build_coleman_forest,
-    build_hyppo_cv_forest,
-    build_hyppo_oob_forest,
+    build_cv_forest,
+    build_oob_forest,
     build_permutation_forest,
 )
 from .monte_carlo import PermutationTest
@@ -16,8 +16,8 @@ __all__ = [
     "PermutationForestClassifier",
     "PermutationForestRegressor",
     "PermutationTest",
-    "build_hyppo_cv_forest",
-    "build_hyppo_oob_forest",
+    "build_cv_forest",
+    "build_oob_forest",
     "build_coleman_forest",
     "build_permutation_forest",
     "PermutationHonestForestClassifier",
