+++ sktree/stats/forestht.py
@@ -1215,7 +1215,8 @@ def _parallel_predict_proba_oob(predict_proba, X, out, idx, test_idx, lock):
 
 
 ForestTestResult = namedtuple(
-    "ForestTestResult", ["observe_test_stat", "permuted_stat", "observe_stat", "pvalue"]
+    "ForestTestResult",
+    ["observe_test_stat", "permuted_stat", "observe_stat", "pvalue", "null_dist"],
 )
 
 
@@ -1266,7 +1267,7 @@ def build_coleman_forest(
     seed : int, optional
         Random seed, by default None.
     return_posteriors : bool, optional
-        Whether or not to return the posteriors, by default False.
+        Whether or not to return the posteriors, by default True.
     **metric_kwargs : dict, optional
         Additional keyword arguments to pass to the metric function.
 
@@ -1283,6 +1284,8 @@ def build_coleman_forest(
     perm_forest_proba : ArrayLike of shape (n_estimators, n_samples, n_outputs)
         The predicted posterior probabilities for each of the permuted estimators
         on their out of bag samples.
+    null_dist : ArrayLike of shape (n_repeats,)
+        The null statistic differences from permuted forests.
 
     References
     ----------
@@ -1290,23 +1293,27 @@ def build_coleman_forest(
     """
     metric_func: Callable[[ArrayLike, ArrayLike], float] = METRIC_FUNCTIONS[metric]
 
-    if covariate_index is None:
-        covariate_index = np.arange(X.shape[1], dtype=int)
-
-    if not isinstance(perm_est, PermutationHonestForestClassifier):
-        raise RuntimeError(
-            f"Permutation forest must be a PermutationHonestForestClassifier, got {type(perm_est)}"
-        )
-
     # build two sets of forests
-    est, orig_forest_proba = build_hyppo_oob_forest(est, X, y, verbose=verbose)
-    perm_est, perm_forest_proba = build_hyppo_oob_forest(
-        perm_est, X, y, verbose=verbose, covariate_index=covariate_index
-    )
+    est, orig_forest_proba = build_oob_forest(est, X, y, verbose=verbose)
+
+    X_null = np.copy(X)
+    y_null = np.copy(y)
+    rng = np.random.default_rng(seed)
+
+    if covariate_index is None:
+        rng.shuffle(y_null)
+    else:
+        temp_col = X_null[:, covariate_index]
+        rng.shuffle(temp_col)
+        X_null[:, covariate_index] = temp_col
+
+    perm_est, perm_forest_proba = build_oob_forest(perm_est, X_null, y_null, verbose=verbose)
 
     # get the number of jobs
     n_jobs = est.n_jobs
 
+    if y.ndim == 1:
+        y = y.reshape(-1, 1)
     metric_star, metric_star_pi = _compute_null_distribution_coleman(
         y,
         orig_forest_proba,
@@ -1325,18 +1332,20 @@ def build_coleman_forest(
 
     # metric^\pi - metric = observed test statistic, which under the
     # null is normally distributed around 0
-    observe_test_stat = permute_stat - observe_stat
+    observe_test_stat = observe_stat - permute_stat
 
     # metric^\pi_j - metric_j, which is centered at 0
     null_dist = metric_star_pi - metric_star
 
     # compute pvalue
     if metric in POSITIVE_METRICS:
-        pvalue = (1 + (null_dist <= observe_test_stat).sum()) / (1 + n_repeats)
-    else:
         pvalue = (1 + (null_dist >= observe_test_stat).sum()) / (1 + n_repeats)
+    else:
+        pvalue = (1 + (null_dist <= observe_test_stat).sum()) / (1 + n_repeats)
 
-    forest_result = ForestTestResult(observe_test_stat, permute_stat, observe_stat, pvalue)
+    forest_result = ForestTestResult(
+        observe_test_stat, permute_stat, observe_stat, pvalue, null_dist
+    )
     if return_posteriors:
         return forest_result, orig_forest_proba, perm_forest_proba, est, perm_est
     else:
@@ -1390,7 +1399,7 @@ def build_permutation_forest(
     seed : int, optional
         Random seed, by default None.
     return_posteriors : bool, optional
-        Whether or not to return the posteriors, by default False.
+        Whether or not to return the posteriors, by default True.
     **metric_kwargs : dict, optional
         Additional keyword arguments to pass to the metric function.
 
@@ -1424,7 +1433,7 @@ def build_permutation_forest(
         )
 
     # train the original forest on unpermuted data
-    est, orig_forest_proba = build_hyppo_oob_forest(est, X, y, verbose=verbose)
+    est, orig_forest_proba = build_oob_forest(est, X, y, verbose=verbose)
     y_pred_proba_orig = np.nanmean(orig_forest_proba, axis=0)
     observe_test_stat = metric_func(y, y_pred_proba_orig, **metric_kwargs)
 
@@ -1443,7 +1452,7 @@ def build_permutation_forest(
         perm_est = clone(perm_est)
         perm_est.set_params(random_state=rng.integers(0, np.iinfo(np.int32).max))
 
-        perm_est, perm_forest_proba = build_hyppo_oob_forest(
+        perm_est, perm_forest_proba = build_oob_forest(
             perm_est, X_perm, y, verbose=verbose, covariate_index=covariate_index
         )
 
@@ -1458,14 +1467,14 @@ def build_permutation_forest(
     else:
         pvalue = (1 + (null_dist <= observe_test_stat).sum()) / (1 + n_repeats)
 
-    forest_result = ForestTestResult(observe_test_stat, permute_stat, None, pvalue)
+    forest_result = ForestTestResult(observe_test_stat, permute_stat, None, pvalue, null_dist)
     if return_posteriors:
         return forest_result, orig_forest_proba, perm_forest_proba
     else:
         return forest_result
 
 
-def build_hyppo_oob_forest(est: ForestClassifier, X, y, verbose=False, **est_kwargs):
+def build_oob_forest(est: ForestClassifier, X, y, verbose=False, **est_kwargs):
     """Build a hypothesis testing forest using oob samples.
 
     Parameters
@@ -1523,7 +1532,7 @@ def build_hyppo_oob_forest(est: ForestClassifier, X, y, verbose=False, **est_kwa
     return est, all_proba
 
 
-def build_hyppo_cv_forest(
+def build_cv_forest(
     est,
     X,
     y,
@@ -1532,7 +1541,7 @@ def build_hyppo_cv_forest(
     verbose=False,
     seed=None,
 ):
-    """Build a hypothesis testing forest using oob samples.
+    """Build a hypothesis testing forest using using cross-validation.
 
     Parameters
     ----------
